refactor: move waitForEvent to @streamr/utils
