+++ packages/broker/test/integration/plugins/websocket/WebsocketPlugin.test.ts
@@ -1,5 +1,6 @@
 import WebSocket from 'ws'
-import { waitForEvent, Queue } from 'streamr-test-utils'
+import { Queue } from 'streamr-test-utils'
+import { waitForEvent } from '@streamr/utils'
 import { Message } from '../../../../src/helpers/PayloadFormat'
 import { createMessagingPluginTest } from '../../createMessagingPluginTest'
