+++ packages/network-tracker/test/integration/tracker-instructions.test.ts
@@ -1,6 +1,7 @@
 import { Tracker } from '../../src/logic/Tracker'
 import { startTracker } from '../../src/startTracker'
-import { runAndWaitForEvents, waitForCondition, waitForEvent } from 'streamr-test-utils'
+import { runAndWaitForEvents, waitForCondition } from 'streamr-test-utils'
+import { waitForEvent } from '@streamr/utils'
 import { InstructionMessage, toStreamID, toStreamPartID } from 'streamr-client-protocol'
 import { NetworkNode, createNetworkNode, NodeEvent } from 'streamr-network'
 import { Event as TrackerServerEvent } from '../../src/protocol/TrackerServer'
