+++ packages/network/test/integration/delivery-of-messages-protocol-layer.test.ts
@@ -11,7 +11,8 @@ import {
     StreamPartIDUtils,
     toStreamID
 } from 'streamr-client-protocol'
-import { runAndWaitForEvents, waitForEvent } from 'streamr-test-utils'
+import { runAndWaitForEvents } from 'streamr-test-utils'
+import { waitForEvent } from '@streamr/utils'
 import { startTracker, Tracker, TrackerServer, TrackerServerEvent } from '@streamr/network-tracker'
 import { Event as NodeToNodeEvent, NodeToNode } from '../../src/protocol/NodeToNode'
 import { Event as NodeToTrackerEvent, NodeToTracker } from '../../src/protocol/NodeToTracker'
