+++ packages/network/test/unit/WebRtcEndpoint.test.ts
@@ -2,8 +2,8 @@ import { MetricsContext } from '../../src/composition'
 import { NodeToTracker } from '../../src/protocol/NodeToTracker'
 import { Tracker, TrackerEvent, startTracker } from '@streamr/network-tracker'
 import { PeerInfo } from '../../src/connection/PeerInfo'
-import { waitForCondition, waitForEvent, runAndWaitForEvents } from 'streamr-test-utils'
-import { wait } from '@streamr/utils'
+import { waitForCondition, runAndWaitForEvents } from 'streamr-test-utils'
+import { wait, waitForEvent } from '@streamr/utils'
 import { Event as EndpointEvent } from '../../src/connection/webrtc/IWebRtcEndpoint'
 import { RtcSignaller } from '../../src/logic/RtcSignaller'
 import { NegotiatedProtocolVersions } from '../../src/connection/NegotiatedProtocolVersions'
