+++ packages/test-utils/package.json
@@ -30,6 +30,7 @@
   "license": "Apache-2.0",
   "dependencies": {
     "@ethersproject/wallet": "^5.5.0",
+    "@streamr/utils": "^0.0.1",
     "cors": "^2.8.5",
     "express": "^4.17.1"
   },
