+++ packages/test-utils/test/utils.test.ts
@@ -2,7 +2,6 @@ import {
     waitForStreamToEnd,
     callbackToPromise,
     toReadableStream,
-    waitForEvent,
     waitForCondition,
     eventsToArray, eventsWithArgsToArray
 } from "../src/utils"
@@ -46,39 +45,6 @@ describe(waitForStreamToEnd, () => {
         waitForStreamToEnd(rs).catch((err) => {
             expect(err).toEqual(new Error('error'))
             done()
-        })
-    })
-})
-
-describe(waitForEvent, () => {
-    it("waits for correct event and records the arguments of invocation", async () => {
-        const emitter = new EventEmitter()
-        setTimeout(() => {
-            emitter.emit("wrongEvent", 666, "beast")
-        }, 0)
-        setTimeout(() => {
-            emitter.emit("correctEvent", 1337, "leet")
-        }, 5)
-        const recordedArgs = await waitForEvent(emitter, "correctEvent")
-        expect(recordedArgs).toEqual([1337, "leet"])
-    })
-
-    it("works on events with zero arguments", async () => {
-        const emitter = new EventEmitter()
-        setTimeout(() => {
-            emitter.emit("wrongEvent", 666, "beast")
-        }, 0)
-        setTimeout(() => {
-            emitter.emit("correctEvent")
-        }, 5)
-        const recordedArgs = await waitForEvent(emitter, "correctEvent")
-        expect(recordedArgs).toEqual([])
-    })
-
-    it("rejects if not event occurs within timeout", async () => {
-        const emitter = new EventEmitter()
-        await waitForEvent(emitter, "correctEvent", 20).catch((err) => {
-            expect(err.message).toEqual("Promise timed out after 20 milliseconds")
         })
     })
 })
