+++ packages/test-utils/tsconfig.jest.json
@@ -6,5 +6,8 @@
     "include": [
         "src/**/*",
         "test/**/*"
+    ],
+    "references": [
+        { "path": "../utils/tsconfig.node.json" }
     ]
 }
