+++ packages/test-utils/tsconfig.node.json
@@ -5,5 +5,8 @@
     },
     "include": [
         "src/**/*"
+    ],
+    "references": [
+        { "path": "../utils/tsconfig.node.json" }
     ]
 }
