+++ packages/utils/src/exports.ts
@@ -6,6 +6,7 @@ import { scheduleAtFixedRate } from './scheduleAtFixedRate'
 import { scheduleAtInterval } from './scheduleAtInterval'
 import { BrandedString } from './types'
 import { wait } from './wait'
+import { waitForEvent } from './waitForEvent'
 import { TimeoutError, withTimeout } from './withTimeout'
 
 export {
@@ -18,5 +19,6 @@ export {
     scheduleAtFixedRate,
     scheduleAtInterval,
     wait,
+    waitForEvent,
     withTimeout,
 }
